Extract cache-transformer
Allows generated operations to be applied to targets other than the cache
