+++ lib/orbit-common/cache.js
@@ -6,6 +6,7 @@ import { eq } from 'orbit/lib/eq';
 import { deprecate } from 'orbit/lib/deprecate';
 import CacheIntegrityProcessor from 'orbit-common/operation-processors/cache-integrity-processor';
 import SchemaConsistencyProcessor from 'orbit-common/operation-processors/schema-consistency-processor';
+import CacheTransformer from './cache-transformer';
 
 /**
  `Cache` provides a thin wrapper over an internally maintained instance of a
@@ -35,21 +36,16 @@ export default Class.extend({
 
     options = options || {};
     var processors = options.processors ? options.processors : [ SchemaConsistencyProcessor, CacheIntegrityProcessor ];
-    this._initProcessors(processors);
 
     // TODO - clean up listener
     this._schema.on('modelRegistered', this._registerModel, this);
-  },
 
-  _initProcessors: function(processors) {
-    this._processors = processors.map(this._initProcessor, this);
-  },
-
-  _initProcessor: function(Processor) {
     var _this = this;
     var retrieve = function(path) { return _this.retrieve(path); };
-    return new Processor(this._schema, retrieve);
+    var transformTarget = function(operation, invert){ return _this._doc.transform(operation, invert); };
+    this._cacheTransformer = new CacheTransformer(retrieve, transformTarget, this._schema, processors);
   },
+
 
   _registerModel: function(model) {
     var modelRootPath = [model];
@@ -61,10 +57,7 @@ export default Class.extend({
   reset: function(data) {
     this._doc.reset(data);
     this._schema.registerAllKeys(data);
-
-    this._processors.forEach(function(processor) {
-      processor.reset(data);
-    });
+    this._cacheTransformer.reset(data);
   },
 
   /**
@@ -141,7 +134,7 @@ export default Class.extend({
    @returns {Array or undefined} Array of inverse operations
    */
   transform: function(operation) {
-    return this._transform(this._normalizeOperation(operation));
+    return this._cacheTransformer.transform(this._normalizeOperation(operation));
   },
 
   _normalizeOperation: function(operation) {
@@ -152,68 +145,5 @@ export default Class.extend({
     }
   },
 
-  _transform: function(operation) {
-    var _this = this;
-    var inverse = [];
-    var op = operation.op;
-    var path = operation.path;
-    var value = operation.value;
-    var currentValue = this.retrieve(path);
-    var relatedOps = [];
-
-    // special case the addition of a `type` collection
-    if (op === 'add' && path.length === 1) {
-      return this._doc.transform(operation, true);
-    }
-
-    if (op === 'add' || op === 'replace') {
-      if (!this.exists(path.slice(0, path.length - 1))) {
-        return;
-      }
-    }
-
-    if (eq(currentValue, value)) return;
-
-    var concatRelatedOps = function(ops) {
-      relatedOps = relatedOps.concat(ops);
-    };
-
-    var performRelatedOps = function() {
-      relatedOps.forEach(function(o) {
-        inverse.push(_this._transform(o));
-      });
-      relatedOps = [];
-    };
-
-    // console.log('Cache#transform', op, path.join('/'), value);
-
-    if (eq(currentValue, value)) return;
-
-    // Query and perform related `before` operations
-    this._processors.forEach(function(processor) {
-      concatRelatedOps(processor.before(operation));
-    });
-    performRelatedOps();
-
-    // Query related `after` operations before performing
-    // the requested operation
-    this._processors.forEach(function(processor) {
-      concatRelatedOps(processor.after(operation));
-    });
-
-    // Perform the requested operation
-    inverse = this._doc.transform(operation, true);
-
-    // Perform related `after` operations after performing
-    // the requested operation
-    performRelatedOps();
-
-    // Query and perform related `finally` operations
-    this._processors.forEach(function(processor) {
-      concatRelatedOps(processor.finally(operation));
-    });
-    performRelatedOps();
-
-    return inverse;
-  }
+  
 });
